Demo Annotations

Ensembl v84 annotations to be used for vignette.

Usage

data(demo_annot)

Format

A 46521 x 6 data frame with gene annotations from Ensembl v84.

id

Ensembl gene IDs

symbol

Associated MGI symbol

chr

Chromosome of the gene

start

Start position (bp) on chromosome

end

End position (bp) on chromosome

strand

What strand the gene is on

Source

Skelly et al., Mapping the Effects of Genetic Variation on Chromatin State and Gene Expression Reveals Loci That Control Ground State Pluripotency, Cell Stem Cell (2020), https://doi.org/10.1016/j.stem.2020.07.005